Improved template builder.
Added check to remove empty tables rows from custom templates as this causes tables to be misplaced within the DomPDF handler.

Added additional placement styles to prevent any divs from being placed where they do not belong.

Added a formatter to the save call to improve the readability of the custom template HTML once it has been saved.

Resolved a bug in DomPDF handler which would cause it to fail when a row does not have a cell present within it.