Billing data consistency
Similar to what already explained in pull request #19.

The billing data with the unserscore (i.e. billing_name) exists just when the order is added.
The correct structure is billing->name, billing->street and so on.

This one works both on order creation and when the pdf invoice generation is called later on passing a MemberOrder instance.